Moving alert configuration/output into its own file
