+++ Gemfile
@@ -7,5 +7,4 @@ gem 'pony'
 gem 'twilio-rb', git: "git://github.com/stevegraham/twilio-rb.git"
 gem 'clockwork'
 
-gem 'curb'
 gem 'safe_yaml'
